Move functions to a helper file
